optimize materialized view refresh
