+++ fe/fe-core/src/main/java/com/starrocks/catalog/MaterializedView.java
@@ -32,7 +32,6 @@ import java.io.DataOutput;
 import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.Collection;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -267,13 +266,6 @@ public class MaterializedView extends OlapTable implements GsonPostProcessable {
         this.refreshScheme = refreshScheme;
     }
 
-    public Set<String> getSyncedPartitionNames(long baseTableId) {
-        return this.getRefreshScheme().getAsyncRefreshContext()
-                .getBaseTableVisibleVersionMap()
-                .computeIfAbsent(baseTableId, k -> Maps.newHashMap())
-                .keySet();
-    }
-
     public Set<String> getNeedRefreshPartitionNames(OlapTable base) {
         Map<String, BasePartitionInfo> baseTableVisibleVersionMap = getRefreshScheme()
                 .getAsyncRefreshContext()
@@ -296,96 +288,6 @@ public class MaterializedView extends OlapTable implements GsonPostProcessable {
             }
         }
         return result;
-    }
-
-    public boolean needRefreshPartition(long baseTableId, Partition baseTablePartition) {
-        BasePartitionInfo basePartitionInfo = this.getRefreshScheme().getAsyncRefreshContext()
-                .getBaseTableVisibleVersionMap()
-                .computeIfAbsent(baseTableId, k -> Maps.newHashMap())
-                .get(baseTablePartition.getName());
-        if (basePartitionInfo == null
-                || basePartitionInfo.getId() != baseTablePartition.getId()
-                || baseTablePartition.getVisibleVersion() > basePartitionInfo.getVersion()) {
-            return true;
-        }
-        return false;
-    }
-
-    public boolean needAddBasePartition(long baseTableId, Partition baseTablePartition) {
-        Map<String, BasePartitionInfo> basePartitionInfoMap = this.getRefreshScheme().getAsyncRefreshContext()
-                .getBaseTableVisibleVersionMap()
-                .computeIfAbsent(baseTableId, k -> Maps.newHashMap());
-        return basePartitionInfoMap.get(baseTablePartition.getName()) == null;
-    }
-
-    public void addBasePartition(long baseTableId, Partition baseTablePartition) {
-        addBasePartition(baseTableId, baseTablePartition, false);
-    }
-
-    public void addBasePartition(long baseTableId, Partition baseTablePartition, boolean isReplay) {
-        Map<String, BasePartitionInfo> basePartitionInfoMap = this.getRefreshScheme().getAsyncRefreshContext()
-                .getBaseTableVisibleVersionMap()
-                .computeIfAbsent(baseTableId, k -> Maps.newHashMap());
-        basePartitionInfoMap.put(baseTablePartition.getName(),
-                new BasePartitionInfo(baseTablePartition.getId(), Partition.PARTITION_INIT_VERSION));
-        if (!isReplay) {
-            GlobalStateMgr.getCurrentState().getEditLog().logAddMvVersionMapInfo(
-                    new MaterializedViewPartitionVersionInfo(this.dbId, this.id, baseTableId, baseTablePartition.getName(),
-                            baseTablePartition.getId(), Partition.PARTITION_INIT_VERSION));
-        }
-    }
-
-    public void updateBasePartition(long baseTableId, Partition baseTablePartition) {
-        updateBasePartition(baseTableId, baseTablePartition, false);
-    }
-
-    public void updateBasePartition(long baseTableId, Partition baseTablePartition, boolean isReplay) {
-        Map<String, BasePartitionInfo> basePartitionInfoMap = this.getRefreshScheme().getAsyncRefreshContext()
-                .getBaseTableVisibleVersionMap()
-                .computeIfAbsent(baseTableId, k -> Maps.newHashMap());
-        basePartitionInfoMap.put(baseTablePartition.getName(),
-                new BasePartitionInfo(baseTablePartition.getId(), baseTablePartition.getVisibleVersion()));
-        if (!isReplay) {
-            GlobalStateMgr.getCurrentState().getEditLog().logAddMvVersionMapInfo(
-                    new MaterializedViewPartitionVersionInfo(this.dbId, this.id, baseTableId, baseTablePartition.getName(),
-                            baseTablePartition.getId(), baseTablePartition.getVisibleVersion()));
-        }
-    }
-
-    public void updateBasePartition(long baseTableId, String basePartitionName, BasePartitionInfo basePartitionInfo) {
-        Map<String, BasePartitionInfo> basePartitionInfoMap = this.getRefreshScheme().getAsyncRefreshContext()
-                .getBaseTableVisibleVersionMap()
-                .computeIfAbsent(baseTableId, k -> Maps.newHashMap());
-        basePartitionInfoMap.put(basePartitionName, basePartitionInfo);
-    }
-
-    public void removeBasePartition(long baseTableId, String baseTablePartitionName, boolean isReplay) {
-        Map<String, BasePartitionInfo> basePartitionInfoMap = this.getRefreshScheme().getAsyncRefreshContext()
-                .getBaseTableVisibleVersionMap()
-                .computeIfAbsent(baseTableId, k -> Maps.newHashMap());
-        basePartitionInfoMap.remove(baseTablePartitionName);
-        if (!isReplay) {
-            GlobalStateMgr.getCurrentState().getEditLog().logRemoveMvVersionMapInfo(
-                    new MaterializedViewPartitionVersionInfo(this.dbId, this.id, baseTableId, baseTablePartitionName, -1, -1));
-        }
-    }
-
-    public int cleanBasePartition(OlapTable base) {
-        int cleanedBasePartitionCounter = 0;
-        Map<String, BasePartitionInfo> basePartitionInfoMap = this.getRefreshScheme().getAsyncRefreshContext()
-                .getBaseTableVisibleVersionMap()
-                .computeIfAbsent(base.getId(), k -> Maps.newHashMap());
-        Iterator<String> basePartitionNameIterator = basePartitionInfoMap.keySet().iterator();
-        while (basePartitionNameIterator.hasNext()) {
-            String basePartitionName = basePartitionNameIterator.next();
-            if (base.getPartition(basePartitionName) == null) {
-                basePartitionNameIterator.remove();
-                GlobalStateMgr.getCurrentState().getEditLog().logRemoveMvVersionMapInfo(
-                        new MaterializedViewPartitionVersionInfo(this.dbId, this.id, base.getId(), basePartitionName, -1, -1));
-                cleanedBasePartitionCounter++;
-            }
-        }
-        return cleanedBasePartitionCounter;
     }
 
     @Override
