+++ fe/fe-core/src/main/java/com/starrocks/persist/EditLog.java
@@ -36,7 +36,6 @@ import com.starrocks.catalog.Database;
 import com.starrocks.catalog.Function;
 import com.starrocks.catalog.FunctionSearchDesc;
 import com.starrocks.catalog.MaterializedView;
-import com.starrocks.catalog.MaterializedViewPartitionVersionInfo;
 import com.starrocks.catalog.MetaVersion;
 import com.starrocks.catalog.Resource;
 import com.starrocks.cluster.Cluster;
@@ -194,16 +193,6 @@ public class EditLog {
                             + " table = " + info.getTableId()
                             + " partitionName = " + info.getPartition().getName());
                     globalStateMgr.replayAddPartition(info);
-                    break;
-                }
-                case OperationType.OP_ADD_MATERIALIZED_VIEW_PARTITION_VERSION_INFO: {
-                    MaterializedViewPartitionVersionInfo info = (MaterializedViewPartitionVersionInfo) journal.getData();
-                    globalStateMgr.replayAddMvPartitionVersionInfo(info);
-                    break;
-                }
-                case OperationType.OP_REMOVE_MATERIALIZED_VIEW_PARTITION_VERSION_INFO: {
-                    MaterializedViewPartitionVersionInfo info = (MaterializedViewPartitionVersionInfo) journal.getData();
-                    globalStateMgr.replayRemoveMvPartitionVersionInfo(info);
                     break;
                 }
                 case OperationType.OP_ADD_PARTITION: {
@@ -1564,15 +1553,8 @@ public class EditLog {
         logEdit(OperationType.OP_DELETE_UNUSED_SHARD, new ShardInfo(shardIds));
     }
 
-    public void logAddMvVersionMapInfo(MaterializedViewPartitionVersionInfo info) {
-        logEdit(OperationType.OP_ADD_MATERIALIZED_VIEW_PARTITION_VERSION_INFO, info);
-    }
-
-    public void logRemoveMvVersionMapInfo(MaterializedViewPartitionVersionInfo info) {
-        logEdit(OperationType.OP_REMOVE_MATERIALIZED_VIEW_PARTITION_VERSION_INFO, info);
-    }
-
     public void logStarMgrOperation(StarMgrJournal journal) {
         logEdit(OperationType.OP_STARMGR, journal);
     }
+
 }
