+++ fe/fe-core/src/main/java/com/starrocks/persist/OperationType.java
@@ -238,8 +238,6 @@ public class OperationType {
     public static final short OP_CREATE_MATERIALIZED_VIEW = 10094;
     public static final short OP_CREATE_INSERT_OVERWRITE = 10095;
     public static final short OP_INSERT_OVERWRITE_STATE_CHANGE = 10096;
-    public static final short OP_ADD_MATERIALIZED_VIEW_PARTITION_VERSION_INFO = 10097;
-    public static final short OP_REMOVE_MATERIALIZED_VIEW_PARTITION_VERSION_INFO = 10098;
 
     // manage system node info 10101 ~ 10120
     public static final short OP_UPDATE_FRONTEND = 10101;
