+++ fe/fe-core/src/main/java/com/starrocks/scheduler/TaskRunBuilder.java
@@ -20,7 +20,7 @@ public class TaskRunBuilder {
         taskRun.setProperties(task.getProperties());
         taskRun.setTask(task);
         if (task.getSource().equals(Constants.TaskSource.MV)) {
-            taskRun.setProcessor(new MvTaskRunProcessor());
+            taskRun.setProcessor(new PartitionBasedMaterializedViewRefreshProcessor());
         } else {
             taskRun.setProcessor(new SqlTaskRunProcessor());
         }
